refactor(webpush): separate endpoint safety policy
